+++ internal/modules/users/usersdomain/usersrepo/user_repository.go
@@ -8,6 +8,6 @@ import (
 )
 
 type UserRepository interface {
-	sharedrepo.GenericRepository[usersdomain.User, uint]
+	sharedrepo.GenericRepository[usersdomain.User, int64]
 	GetByEmail(ctx context.Context, email string) (*usersdomain.User, error)
 }
